+++ src/main/java/de/yuga/spacebattle/backend/services/MasterOfTheUniverseService.java
@@ -228,12 +228,34 @@ public class MasterOfTheUniverseService {
     public void transform() {
         validateUniverse();
         LOGGER.info("---------------------------- transforming the universe ----------------------------");
-        final boolean transformationNeeded = false;
+        final boolean transformationNeeded = true;
         if (transformationNeeded) {
+            cleanUpUnusedAmmunitionFittings(); /* fixme remove me after release */
+
             LOGGER.info("---------------------------- done transforming -------------------------------");
         } else {
             LOGGER.info("---------------------------- nothing to transform ----------------------------");
         }
+    }
+
+    private void cleanUpUnusedAmmunitionFittings() {
+        final List<ShipClass> shipClasses = shipClassService.findAll();
+        final List<ShipClass> toStore = new ArrayList<>();
+        shipClasses.forEach(shipClass -> {
+            final Set<Missile> allowedMissiles = shipClass.getFittings().stream()
+                    .map(AlignedFitting::getLauncher)
+                    .filter(Objects::nonNull)
+                    .map(Launcher::getAllowedMissiles)
+                    .flatMap(Collection::stream)
+                    .collect(Collectors.toSet());
+
+            final List<AmmunitionFitting> unusedMissiles = shipClass.getAmmunitionFittings().stream().filter(af -> !allowedMissiles.contains(af.getMissile())).collect(Collectors.toList());
+            if (!unusedMissiles.isEmpty()) {
+                unusedMissiles.forEach(shipClass.getAmmunitionFittings()::remove);
+                toStore.add(shipClass);
+            }
+        });
+        shipClassService.saveAll(toStore);
     }
 
     @SuppressWarnings("unused")